UnityEvents y HUD: Imagen y Texto Inicial

+++ Proyecto Unity 2D/Assets/Scripts/Jugador/PerfilJugador.cs
@@ -40,8 +40,8 @@ public class PerfilJugador : ScriptableObject
     [Header("Configuraciones de Atributos")]
     [SerializeField]
     [Range(5, 10)]
-    private float vida = 5f;
-    public float Vida { get => vida; set => vida = value; }
+    private int vida = 5;
+    public int Vida { get => vida; set => vida = value; }
 
 
     [Header("Configuraciones SFX")]
